# Faster conversion from all codes (not country names)